restructure data access dir
