+++ src/main/java/data_access/database/DatabaseAccessInterface.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.database;
 
 public interface DatabaseAccessInterface {
