+++ src/main/java/data_access/database/PostGreSQLAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.database;
 
 import java.io.BufferedReader;
 import java.io.File;
