+++ src/main/java/data_access/image/ImageCacheAccessInterface.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.image;
 
 import java.awt.Image;
 import java.awt.image.BufferedImage;
