+++ src/main/java/data_access/image/ImageDataAccessInterface.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.image;
 
 import java.awt.image.BufferedImage;
