+++ src/main/java/data_access/pokemon/PokemonCardDataAccessInterface.java
@@ -1,13 +1,12 @@
-package data_access;
+package data_access.pokemon;
 
 import java.awt.Image;
 import java.util.ArrayList;
 
 import entity.Card;
 
-public interface PokemonGuruDataAccessInterface {
+public interface PokemonCardDataAccessInterface {
     public Card getCard(String id);
-    public Image getImage(String url);
     public ArrayList<Card> searchCards(PokemonGuruCardSearchFilter filter);
     public ArrayList<Card> searchCards(String query);
 }
