+++ src/main/java/data_access/pokemon/PokemonGuruDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.pokemon;
 
 import java.awt.Image;
 import java.io.BufferedReader;
