+++ src/main/java/usecase/login/LoginDataAccessInterface.java
@@ -1,0 +1,7 @@
+package usecase.login;
+
+import entity.User;
+
+public interface LoginDataAccessInterface {
+    public User getUser(String username, String password);
+}
